requirements.txt
asgiref
Django
mysqlclient
Pillow
sqlparse
tzdata
# django-tinymce